updated file structure: changed name of src to client, cleaned out Grid file, added algorithm folder for dfs, added data folder for creating Adj list and matrix
