+++ client/app.js
@@ -5,6 +5,6 @@ import {Provider} from "react-redux";
 import MainComponent from "./components/Main.js";
 
 ReactDOM.render(
-                <MainComponent />
-                ,document.getElementById("main")
-                );
+  <MainComponent />
+  ,document.getElementById("main")
+);
