+++ client/components/Main.js
@@ -1,5 +1,4 @@
 import React from "react";
-import ReactDOM from "react-dom"
 import Grid from './Grid.js'
 
 export default class Main extends React.Component{
